requirements.txt
streamlit==1.37.0
PyPDF2==3.0.1
openai==0.28.0
faiss-cpu==1.8.0.post1
fastapi==0.112.0
uvicorn==0.30.5
pydantic==2.8.2
PyMuPDF==1.24.9